IconToggle: abort onPress if the user scrolls (#392)
* IconToggle: abort onPress if the user scrolls

TouchableWithoutFeedback's onPressIn and onPressOut props are not cancelled when a user scrolls. IconToggle was previously calling its onPress prop in a handler for TouchableWithoutFeedback's onPressOut prop. As a result, the consumer-specified onPress handler would not abort if the user initiated scroll events.

This commit moves IconToggle's onPress logic to a TouchableWithoutFeedbacks's onPress handler, causing it to abort if a user scrolls.

* IconToggle: fix linting

Removes unused onPress definition from onPressOut and modfiies onPress to accomodoate `no-unused-expressions`.